fix: deduplicate migration cleanup, add error handling, remove dead code
- Extract removeOldGeneratedTypes into migration.ts to avoid duplication
  and circular dependency between index.ts and serving/generator.ts
- Add try-catch to resolveEndpointsFromServerFile so AST parse failures
  fall back gracefully to env var
- Remove unused root variable in analytics vite-plugin
- Clean up template tsconfig.server.json: remove dead emit options
  (outDir, declaration, declarationMap, sourceMap) since noEmit is set
